Moving CoreConfig to new file
